Revert "String caps change" This reverts commit d39265e7e70dcfb5da9d5dcd833c849472f7cc4d.
The capfirst call at this place forces the "lazy translation string" to be translated immediately when the code is executed (and using english as the default language).
This aint't right - it needs to be kept as a lazy translation string. It is only translated to a string in the template where we display to form to the user in the right language.